Added Bambenek feeds support
"Bambenek Consulting is not the standard run-of-the-mill consultancy. Built by information security expert, John Bambenek, it’s a firm whose services revolve around security and cybercrime. Many people can run a mail server, but not everyone can secure it. For organizations looking for to put a premium on security in the management of their IT assets, we’re here to assist."